download: merge integration test files
